feat(hydro_lang)!: don't pull in dfir_rs during the compilation stage
Because `hydro_lang` is responsible for _generating_ DFIR code, it doesn't actually need to depend on the runtime (`dfir_rs`), other than when it is used in the (legacy) macro mode or when we want to include utilities for runtime logic (`resource_measurement`). This sticks those pieces under feature flags and makes `dfir_rs` an optional dependency, which reduces the compile tree for crates like `hydro_test`.